+++ src/MetricsTest.php
@@ -92,4 +92,31 @@ final class MetricsTest extends TestCase
 
         self::assertSame(1_000_000_000.0, Metrics::fromSnapshots($start, $end)->cpuTime);
     }
+
+    #[Test]
+    public function it_can_returns_the_metrics_in_a_human_readable_format(): void
+    {
+        $start = new Snapshot(new DateTimeImmutable(), hrtime(true), [
+            'ru_utime.tv_sec'  => 1,
+            'ru_utime.tv_usec' => 500_000, // 1.5s user
+            'ru_stime.tv_sec'  => 0,
+            'ru_stime.tv_usec' => 250_000, // 0.25s system
+        ], 1000, 2000, 3000, 4000);
+        $end = new Snapshot(new DateTimeImmutable(), hrtime(true) + 1, [
+            'ru_utime.tv_sec'  => 2,
+            'ru_utime.tv_usec' => 0,       // 2.0s user
+            'ru_stime.tv_sec'  => 0,
+            'ru_stime.tv_usec' => 750_000, // 0.75s system
+        ], 1100, 2100, 3100, 4100);
+        $metrics = Metrics::fromSnapshots($start, $end);
+
+        $humans = $metrics->forHuman();
+        self::assertIsArray($humans);
+        self::assertArrayHasKey('cpu_time', $humans);
+        self::assertArrayHasKey('memory_usage', $humans);
+        self::assertIsString($metrics->forHuman('execution_time'));
+
+        $this->expectException(InvalidArgument::class);
+        $metrics->forHuman('foobar');
+    }
 }