+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 All Notable changes to `bakame/aide-profiler` will be documented in this file.
 
+## [Next](https://github.com/bakame-php/aide-profiler/compare/0.8,0...main) - TBD
+
+### Added
+
+- Added `Metrics::forHuman` to ease getting human-readable metrics representations.
+
+### Fixed
+
+- None
+
+### Deprecated
+
+- None
+
+### Removed
+
+- None
+
 ## [0.8.0 - Harare](https://github.com/bakame-php/aide-profiler/compare/0.7.1...0.8.0) - 2025-07-07
 
 ### Added